Make gradle able to build robot release jar
Upgraded the build.gradle file to make the jar target create a
robot jar file that robocode can use and that can be deployed
in the rumble.

Now what's missing is a working build target that copies the
Kotlin sources as well to work around a robocode class loading
issue.